block processing test into own package to avoid cycle
