+++ beacon-chain/blockchain/fork_choice_test.go
@@ -31,7 +31,7 @@ func generateTestGenesisStateAndBlock(
 			Pubkey: pubkey,
 		}
 		balance := params.BeaconConfig().MaxDepositAmount
-		depositData, err := b.EncodeDepositData(depositInput, balance, time.Now().Unix())
+		depositData, err := helpers.EncodeDepositData(depositInput, balance, time.Now().Unix())
 		if err != nil {
 			t.Fatalf("Could not encode deposit: %v", err)
 		}
