+++ beacon-chain/blockchain/service_test.go
@@ -9,6 +9,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	gethTypes "github.com/ethereum/go-ethereum/core/types"
@@ -124,7 +126,7 @@ func setupInitialDeposits(t *testing.T) []*pb.Deposit {
 func createPreChainStartDeposit(t *testing.T, pk []byte) *pb.Deposit {
 	depositInput := &pb.DepositInput{Pubkey: pk}
 	balance := params.BeaconConfig().MaxDepositAmount
-	depositData, err := b.EncodeDepositData(depositInput, balance, time.Now().Unix())
+	depositData, err := helpers.EncodeDepositData(depositInput, balance, time.Now().Unix())
 	if err != nil {
 		t.Fatalf("Cannot encode data: %v", err)
 	}
