+++ beacon-chain/db/BUILD.bazel
@@ -45,7 +45,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/attestations:go_default_library",
-        "//beacon-chain/core/blocks:go_default_library",
+        "//beacon-chain/core/helpers:go_default_library",
         "//beacon-chain/core/validators:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/hashutil:go_default_library",
